+++ packages/contracts/test/mocks/LocalProofE2E.sol
@@ -22,6 +22,7 @@ contract LocalProofE2EVerifier is IRiscZeroVerifier {
 contract LocalProofE2EStateOracle is IBaseAnalysisStateOracle {
     mapping(uint64 blockNumber => bytes32 blockHash) public canonicalBlockHashes;
     bool public historicalBackfillStarted;
+    bool public historicalCoverageComplete;
 
     function setCanonical(uint64 blockNumber, bytes32 blockHash) external {
         canonicalBlockHashes[blockNumber] = blockHash;